pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openagent-framework"
version = "0.2.11"
description = "Simplified LLM agent framework with MCP, memory, and multi-channel support"
readme = "docs/README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "mcp>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18.6",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "websockets>=14.1",
    "aiosqlite>=0.20.0",
    "croniter>=2.0.0",
    "claude-agent-sdk>=0.1.50",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3.0"]
whatsapp = ["whatsapp-api-client-python>=0.0.46"]
# Local voice transcription via faster-whisper (CTranslate2). Runs offline
# on CPU, no API key needed. First call downloads the model (~150 MB for
# the default 'base'). Set OPENAGENT_WHISPER_MODEL to pick a different
# size (tiny, base, small, medium, large-v3).
voice = ["faster-whisper>=1.0.0"]
all = [
    "python-telegram-bot>=21.0",
    "discord.py>=2.3.0",
    "whatsapp-api-client-python>=0.0.46",
    "faster-whisper>=1.0.0",
]

[project.scripts]
openagent = "openagent.cli:main"

[tool.setuptools.packages.find]
include = ["openagent*"]

[tool.setuptools.package-data]
openagent = [
    "mcps/**/*.ts",
    "mcps/**/*.json",
    "mcps/**/*.py",
    "mcps/**/*.txt",
    "mcps/**/.gitignore",
]
